+++ data/musics.json
@@ -1,1 +1,1 @@
-[{"publish_date":"2018-02-04","on_spotify":true},{"publish_date":"2008-06-12","on_spotify":false},{"publish_date":"1999-02-11","on_spotify":true}]
+[{"publish_date":"2018-02-04","on_spotify":true},{"publish_date":"2008-06-12","on_spotify":false},{"publish_date":"1999-02-11","on_spotify":true},{"publish_date":"2012-10-01","on_spotify":true}]
