+++ modules/music_album.rb
